+++ NovaLibrary/src/Config.h
@@ -94,7 +94,7 @@ public:
     int GetSaveFreq();
     int GetTcpCheckFreq();
     int GetTcpTimout();
-    int GetThinningDistance();
+    double GetThinningDistance();
 
     double GetClassificationThreshold();
     double GetSaSleepDuration();
@@ -129,7 +129,7 @@ public:
     void SetSaveFreq(int saveFreq);
     void SetTcpCheckFreq(int tcpCheckFreq);
     void SetTcpTimout(int tcpTimout);
-    void SetThinningDistance(int thinningDistance);
+    void SetThinningDistance(double thinningDistance);
     void SetUseTerminals(bool useTerminals);
     void SetKey(std::string key);
     void SetNeigbors(std::vector<in_addr_t> neighbors);
@@ -201,7 +201,7 @@ private:
 	int m_classificationTimeout;
 	int m_saPort;
 	int m_k;
-	int m_thinningDistance;
+	double m_thinningDistance;
 	int m_saveFreq;
 	int m_dataTTL;
 	int m_saMaxAttempts;
